feat: Post detail get 로직 구현(#17)
- user 객체 가져오기 로직 수정 완료

+++ src/main/java/com/develop_ping/union/post/domain/service/PostServiceImpl.java
@@ -1,10 +1,13 @@
 package com.develop_ping.union.post.domain.service;
 
 import com.develop_ping.union.post.domain.PostManager;
-import com.develop_ping.union.post.domain.dto.command.*;
-import com.develop_ping.union.post.domain.dto.info.*;
+import com.develop_ping.union.post.domain.dto.command.PostCreationCommand;
+import com.develop_ping.union.post.domain.dto.command.PostUpdateCommand;
+import com.develop_ping.union.post.domain.dto.info.PostInfo;
 import com.develop_ping.union.post.domain.entity.Post;
-import com.develop_ping.union.user.domain.User;
+import com.develop_ping.union.post.exception.PostPermissionDeniedException;
+import com.develop_ping.union.user.domain.UserManager;
+import com.develop_ping.union.user.domain.entity.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -14,41 +17,40 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class PostServiceImpl implements PostService {
     private final PostManager postManager;
+    private final UserManager userManager;
 
     @Override
-    public PostCreationInfo createPost(PostCreationCommand command) {
+    public PostInfo createPost(PostCreationCommand command) {
         log.info("[ CALL: PostService.createPost() ] title: {}",command.getTitle());
 
-        // TODO: token에서 유저 추출하기
-        User user = User.builder().id(1L).build();
-
-        // TODO: photo List 저장하기 -> 이거 따로 빼기로 함!!!
+        User user = userManager.findByToken(command.getToken());
+//        User user = userManager.findById(1L);
 
         // TODO: thumbnail에 들어갈 사진 압축하기
         String thumbnail = null;
-        if (command.getPhotos() != null && !command.getPhotos().isEmpty()) {
-            thumbnail = command.getPhotos().get(0);
-            // 압축 메소드 호출? (미구현)
+        if (command.getThumbnail() != null) {
+            // thumbnail을 compress하는 메소드 호출
         }
 
         Post post = postManager.save(Post.of(command, user, thumbnail));
-        return PostCreationInfo.of(post);
+        return PostInfo.of(post);
     }
 
     @Override
-    public PostUpdateInfo updatePost(PostUpdateCommand command) {
+    public PostInfo updatePost(PostUpdateCommand command) {
         log.info("[ CALL: PostService.updatePost() ] post id: {}", command.getId());
 
         Post post = postManager.findById(command.getId());
+        User user = userManager.findByToken(command.getToken());
+//        User user = userManager.findById(1L);
 
-        // TODO: command.getToken()으로 사용자 권한 확인하는 로직 추가
+        validatePostOwner(user, post);
 
         post.updateTitle(command.getTitle());
         post.updateContent(command.getContent());
 
         Post updatedPost = postManager.save(post);
-
-        return PostUpdateInfo.of(updatedPost);
+        return PostInfo.of(updatedPost);
     }
 
     @Override
@@ -56,9 +58,29 @@ public class PostServiceImpl implements PostService {
         log.info("[ CALL: PostService.deletePost() ] post id: {}", postId);
 
         Post post = postManager.findById(postId);
+        User user = userManager.findByToken(token);
+//        User user = userManager.findById(1L);
 
-        // TODO: command.getToken()으로 사용자 권한 확인하는 로직 추가
+        validatePostOwner(user, post);
 
         postManager.delete(post);
+    }
+
+    @Override
+    public PostInfo getPost(Long postId) {
+        log.info("[ CALL: PostService.getPost() ] post id: {}", postId);
+
+        Post post = postManager.findById(postId);
+
+        post.incrementViews();
+        postManager.save(post);
+
+        return PostInfo.of(post);
+    }
+
+    private void validatePostOwner(User user, Post post) {
+        if (!user.equals(post.getUser())) {
+            throw new PostPermissionDeniedException(user.getId(), post.getId());
+        }
     }
 }
